requirements.txt
torch
torchvision
pandas
numpy
scikit-learn
matplotlib
seaborn
pillow
tqdm
kaggle
